+++ lib/weber/handler/weber_req_handler_result.ex
@@ -25,7 +25,12 @@ defmodule Handler.WeberReqHandler.Result do
                  |> String.downcase
                  |> :erlang.binary_to_list
     {:ok, file_content} = File.read(:lists.nth(1, find_file_path(get_all_files(app.views), filename)))
-    Weber.Helper.ContentFor.content_for(:layout)
+
+    case :lists.keyfind('__layout__', 1, app.controller.__info__(:functions)) do
+      false -> :ok;
+      _ -> Weber.Helper.ContentFor.content_for(:layout, file_content, app.controller.__layout__)
+    end
+
     {:render, 200, (EEx.eval_string add_helpers_imports(file_content), data), [{"Content-Type", "text/html"} | headers]}
   end